fix: update the template to my personal data
