+++ .env
@@ -1,2 +1,2 @@
 NEXT_PUBLIC_WEBSITE_URL=https://hamishw.com
-NEXT_PUBLIC_API_URL=https://api.hamishw.com
+NEXT_PUBLIC_API_URL=https://api.mailjet.com/v3.1
