+++ .storybook/manager-head.html
@@ -4,14 +4,14 @@
   }
 </style>
 
-<title>Hamish Williams | Storybook</title>
+<title>João Emerson | Storybook</title>
 <meta
   name="description"
   content="My custom design component system created from the ground up used in building my portfolio"
 />
 
 <meta name="theme-color" content="#111111" />
-<meta name="author" content="Hamish Williams" />
+<meta name="author" content="João Emerson" />
 
 <link rel="manifest" href="https://hamishw.com/manifest.json" />
 <link rel="shortcut icon" href="https://hamishw.com/favicon.svg" />
@@ -24,7 +24,7 @@
 <meta property="og:image:height" content="675" />
 <meta property="og:image:alt" content="Banner for the site" />
 
-<meta property="og:title" content="Hamish Williams | Storybook" />
+<meta property="og:title" content="João Emerson | Storybook" />
 <meta property="og:site_name" content="hamishw.com" />
 <meta property="og:url" content="https://storybook.hamishw.com" />
 <meta
@@ -37,7 +37,7 @@
   name="twitter:description"
   content="My custom design component system created from the ground up used in building my portfolio"
 />
-<meta name="twitter:title" content="Hamish Williams | Storybook" />
+<meta name="twitter:title" content="João Emerson | Storybook" />
 <meta name="twitter:site" content="@hamishMW" />
 <meta name="twitter:image" content="https://hamishw.com/social-image.png" />
