+++ .storybook/manager.js
@@ -5,7 +5,7 @@ addons.setConfig({
   theme: {
     ...themes.dark,
     brandImage: 'https://hamishw.com/icon.svg',
-    brandTitle: 'Hamish Williams Components',
+    brandTitle: 'João Emerson Components',
     brandUrl: 'https://hamishw.com',
   },
 });
