--- public/humans.txt
@@ -1,38 +1,0 @@
-__  __  __
-\ \ \ \ \/
- \ \/\ \
-  \/  \/
-
-The humans.txt file explains the team, technology,
-and graphic assets behind this site humanstxt.org.
-
-_______________________________________________________________________________
-
-DESIGNER | デザイナー
-
-Hamish Williams
-Hand crafted design & web development
-
-Design tools:
-Figma
-Photoshop
-Blender
-
-twitter.com/hamishmw
-github.com/hamishmw
-
-_______________________________________________________________________________
-
-TECHNOLOGY | テクノロジー
-
-React
-Next.js
-Three.js
-Framer Motion
-
-_______________________________________________________________________________
-
-FONTS | フォント
-
-Gotham
-Hiragino Kaku Gothic
