+++ public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "short_name": "Hamish Williams",
-  "name": "Design Portfolio of Hamish Williams",
+  "short_name": "João Emerson",
+  "name": "Design Portfolio of João Emerson",
   "icons": [
     {
       "src": "favicon.png",
