+++ src/components/Footer/Footer.js
@@ -1,4 +1,4 @@
-import { Link } from 'components/Link';
+
 import { Text } from 'components/Text';
 import { classes } from 'utils/style';
 import styles from './Footer.module.css';
@@ -7,11 +7,8 @@ export const Footer = ({ className }) => (
   <footer className={classes(styles.footer, className)}>
     <Text size="s" align="center">
       <span className={styles.date}>
-        {`© ${new Date().getFullYear()} Hamish Williams.`}
+        {`© ${new Date().getFullYear()} João Emerson. Inspired by Hamish Williams`}
       </span>
-      <Link secondary className={styles.link} href="/humans.txt" target="_self">
-        Crafted by yours truly
-      </Link>
     </Text>
   </footer>
 );
