+++ src/components/Icon/Icon.js
@@ -15,6 +15,7 @@ import Pause from './svg/pause.svg';
 import Play from './svg/play.svg';
 import Send from './svg/send.svg';
 import Twitter from './svg/twitter.svg';
+import Linkedin from './svg/linkedin.svg';
 
 export const icons = {
   arrowLeft: ArrowLeft,
@@ -32,6 +33,7 @@ export const icons = {
   play: Play,
   send: Send,
   twitter: Twitter,
+  linkedin: Linkedin,
 };
 
 export const Icon = ({ icon, className, ...rest }) => {
