+++ src/components/Meta/Meta.js
@@ -1,7 +1,7 @@
 import Head from 'next/head';
 
 const siteUrl = process.env.NEXT_PUBLIC_WEBSITE_URL;
-const name = 'Hamish Williams';
+const name = 'João Emerson';
 const twitterHandle = '@hamishMW';
 const defaultOgImage = `${siteUrl}/social-image.png`;
