+++ src/components/Monogram/Monogram.js
@@ -18,7 +18,7 @@ export const Monogram = forwardRef(({ highlight, className, ...props }, ref) => 
     >
       <defs>
         <clipPath id={clipId}>
-          <path d="M16.525 28.462l7.18-18.35.003-.001 9.72 18.442a.838.838 0 001.524-.093l3.39-8.824a.846.846 0 00-.04-.686L30.307 3.605A6.698 6.698 0 0024.367 0h-4.6a.84.84 0 00-.74 1.23l3.63 6.887-3.655 9.15-7.12-13.662A6.698 6.698 0 005.942 0h-4.6a.842.842 0 00-.748 1.23L15 28.554a.839.839 0 001.524-.092zM42.392 8.806a.835.835 0 00.387-.446v.001l2.67-7.23a.838.838 0 00-.785-1.129h-6.578a.837.837 0 00-.736 1.238l3.907 7.226c.22.41.729.56 1.135.34z" />
+          <path d="M10.4895 3.12V18C10.4895 20.3467 10.2015 22.2667 9.6255 23.76C9.07083 25.2533 8.39883 26.416 7.6095 27.248C6.79883 28.08 5.95617 28.656 5.0815 28.976C4.22817 29.3173 3.49217 29.488 2.8735 29.488C2.12683 29.488 1.5295 29.3173 1.0815 28.976C0.612167 28.656 0.3775 28.1547 0.3775 27.472C0.3775 26.8747 0.558833 26.416 0.9215 26.096C1.26283 25.7973 1.73217 25.648 2.3295 25.648C2.7135 25.648 3.04417 25.7227 3.3215 25.872C3.59883 26.0427 3.8335 26.288 4.0255 26.608C4.13217 26.8427 4.23883 26.992 4.3455 27.056C4.4735 27.12 4.6335 27.152 4.8255 27.152C5.12417 27.152 5.41217 26.992 5.6895 26.672C5.98817 26.3733 6.25483 25.84 6.4895 25.072C6.70283 24.304 6.8735 23.2693 7.0015 21.968C7.15083 20.6453 7.2255 18.9707 7.2255 16.944V3.12C7.2255 2.672 7.1615 2.34133 7.0335 2.128C6.92683 1.91467 6.79883 1.776 6.6495 1.712C6.41483 1.648 6.14817 1.584 5.8495 1.52C5.57217 1.456 5.3055 1.41333 5.0495 1.392C4.7935 1.392 4.6015 1.36 4.4735 1.296C4.36683 1.21067 4.3135 1.01867 4.3135 0.719999C4.3135 0.570665 4.32417 0.463999 4.3455 0.4C4.38817 0.314666 4.4415 0.261333 4.5055 0.24C4.5695 0.197332 4.65483 0.175999 4.7615 0.175999C4.86817 0.175999 4.99617 0.175999 5.1455 0.175999C5.18817 0.175999 5.3695 0.186665 5.6895 0.207998C6.0095 0.207998 6.37217 0.207998 6.7775 0.207998C7.18283 0.229333 7.5775 0.24 7.9615 0.24C8.3455 0.24 8.64417 0.24 8.8575 0.24C9.0495 0.24 9.3375 0.24 9.7215 0.24C10.1055 0.24 10.5002 0.229333 10.9055 0.207998C11.3108 0.207998 11.6735 0.207998 11.9935 0.207998C12.3135 0.186665 12.5055 0.175999 12.5695 0.175999C12.6975 0.175999 12.8148 0.175999 12.9215 0.175999C13.0282 0.175999 13.1135 0.197332 13.1775 0.24C13.2415 0.261333 13.2948 0.314666 13.3375 0.4C13.3802 0.463999 13.4015 0.570665 13.4015 0.719999C13.4015 1.01867 13.3375 1.21067 13.2095 1.296C13.1028 1.36 12.9215 1.392 12.6655 1.392C12.4095 1.41333 12.1322 1.456 11.8335 1.52C11.5348 1.584 11.2788 1.648 11.0655 1.712C10.8948 1.776 10.7562 1.91467 10.6495 2.128C10.5428 2.34133 10.4895 2.672 10.4895 3.12ZM30.5343 5.872C30.3209 5.936 30.1396 5.95733 29.9903 5.936C29.8623 5.89333 29.7663 5.744 29.7023 5.488C29.5743 5.14667 29.4249 4.784 29.2543 4.4C29.1049 4.016 28.9343 3.65333 28.7423 3.312C28.5503 2.94933 28.2836 2.66133 27.9423 2.448C27.6223 2.23467 27.1956 2.128 26.6623 2.128H23.8783C23.6223 2.128 23.3556 2.13867 23.0783 2.16C22.8223 2.16 22.5769 2.192 22.3423 2.256C22.1503 2.32 22.0116 2.46933 21.9263 2.704C21.8623 2.91733 21.8303 3.22667 21.8303 3.632V10.096C21.8303 10.2453 21.8516 10.3413 21.8943 10.384C21.9583 10.4267 22.0649 10.448 22.2143 10.448H25.8623C26.5023 10.448 26.8969 10.3947 27.0463 10.288C27.2169 10.16 27.3663 9.95733 27.4943 9.68C27.5583 9.488 27.6329 9.232 27.7183 8.912C27.8036 8.592 27.8996 8.176 28.0063 7.664C28.0276 7.472 28.0809 7.35467 28.1663 7.312C28.2729 7.248 28.4436 7.22667 28.6783 7.248C28.9129 7.29067 29.0516 7.35467 29.0943 7.44C29.1583 7.52533 29.1903 7.664 29.1903 7.856C29.1263 8.496 29.0836 9.168 29.0623 9.872C29.0409 10.5547 29.0303 11.088 29.0303 11.472C29.0303 11.7707 29.0409 12.304 29.0623 13.072C29.0836 13.8187 29.1263 14.512 29.1903 15.152C29.1903 15.344 29.1583 15.4933 29.0943 15.6C29.0516 15.6853 28.9129 15.7387 28.6783 15.76C28.4436 15.8027 28.2729 15.792 28.1663 15.728C28.0809 15.6427 28.0276 15.4933 28.0063 15.28C27.9209 14.832 27.8249 14.4053 27.7183 14C27.6329 13.5733 27.5263 13.2 27.3983 12.88C27.3129 12.624 27.1849 12.4427 27.0143 12.336C26.8436 12.208 26.4383 12.144 25.7983 12.144H22.2143C22.0649 12.144 21.9583 12.1653 21.8943 12.208C21.8516 12.2507 21.8303 12.336 21.8303 12.464V20.528C21.8303 20.9547 21.8623 21.2747 21.9263 21.488C21.9903 21.7013 22.1183 21.8293 22.3103 21.872C22.5449 21.936 22.8009 21.9787 23.0783 22C23.3556 22 23.6223 22 23.8783 22H28.0063C28.5396 22 28.9769 21.8933 29.3183 21.68C29.6596 21.4667 29.9369 21.168 30.1503 20.784C30.3849 20.336 30.6089 19.8453 30.8223 19.312C31.0356 18.7787 31.2383 18.256 31.4303 17.744C31.5156 17.5093 31.6009 17.3813 31.6863 17.36C31.7929 17.3173 31.9529 17.3173 32.1663 17.36C32.2943 17.4027 32.3796 17.4453 32.4223 17.488C32.4863 17.5307 32.5289 17.584 32.5503 17.648C32.5716 17.712 32.5716 17.7867 32.5503 17.872C32.5503 17.936 32.5503 18.0107 32.5503 18.096C32.4223 18.9707 32.2623 19.888 32.0703 20.848C31.8783 21.7867 31.6756 22.672 31.4623 23.504C31.3983 23.6747 31.3343 23.7707 31.2703 23.792C31.2063 23.8347 31.1103 23.856 30.9823 23.856C30.7903 23.856 30.3743 23.856 29.7343 23.856C29.1156 23.856 28.4329 23.8453 27.6863 23.824C26.9183 23.8027 26.1716 23.7813 25.4463 23.76C24.7209 23.76 24.1556 23.76 23.7503 23.76C23.2809 23.76 22.6409 23.76 21.8303 23.76C21.0409 23.7813 20.2623 23.8027 19.4943 23.824C18.7049 23.8453 18.0009 23.856 17.3823 23.856C16.7636 23.856 16.4116 23.856 16.3263 23.856C16.1983 23.856 16.0809 23.856 15.9743 23.856C15.8676 23.856 15.7716 23.8453 15.6863 23.824C15.6009 23.7813 15.5369 23.7173 15.4943 23.632C15.4516 23.568 15.4303 23.4507 15.4303 23.28C15.4303 23.152 15.4409 23.0453 15.4622 22.96C15.5049 22.8747 15.5583 22.8107 15.6223 22.768C15.6863 22.7467 15.7609 22.7253 15.8463 22.704C15.9529 22.6827 16.0809 22.6613 16.2303 22.64C16.5503 22.6187 16.8489 22.576 17.1263 22.512C17.4036 22.448 17.6809 22.352 17.9583 22.224C18.1716 22.1387 18.3209 21.9787 18.4062 21.744C18.5129 21.488 18.5663 21.2 18.5663 20.88V3.184C18.5663 2.82133 18.5129 2.52267 18.4062 2.288C18.3209 2.05333 18.1716 1.89333 17.9583 1.808C17.8089 1.76533 17.6596 1.73333 17.5103 1.712C17.3609 1.66933 17.2116 1.63733 17.0623 1.616C16.8916 1.59467 16.7209 1.57333 16.5503 1.552C16.4009 1.53067 16.2516 1.50933 16.1023 1.488C15.8249 1.488 15.6223 1.456 15.4943 1.392C15.3876 1.328 15.3343 1.136 15.3343 0.815998C15.3343 0.666665 15.3556 0.559999 15.3983 0.495998C15.4409 0.410665 15.4943 0.357332 15.5583 0.335999C15.6436 0.293332 15.7396 0.271999 15.8463 0.271999C15.9529 0.271999 16.0703 0.271999 16.1983 0.271999C16.2836 0.271999 16.6249 0.282666 17.2223 0.303999C17.8409 0.303999 18.5343 0.314665 19.3023 0.335999C20.0703 0.357332 20.8276 0.367998 21.5743 0.367998C22.3423 0.367998 22.9503 0.367998 23.3983 0.367998C23.8036 0.367998 24.3476 0.367998 25.0303 0.367998C25.7129 0.367998 26.3849 0.357332 27.0463 0.335999C27.7076 0.314665 28.2836 0.303999 28.7743 0.303999C29.2649 0.282666 29.5316 0.271999 29.5743 0.271999C29.7236 0.271999 29.8409 0.293332 29.9263 0.335999C30.0329 0.378666 30.1076 0.496 30.1503 0.688C30.2783 1.47733 30.4063 2.23467 30.5343 2.96C30.6623 3.68533 30.7903 4.42133 30.9183 5.168C30.9396 5.36 30.9289 5.52 30.8863 5.648C30.8436 5.75467 30.7263 5.82933 30.5343 5.872Z" fill="black"/>
         </clipPath>
       </defs>
       <rect clipPath={`url(#${clipId})`} width="100%" height="100%" />
