+++ src/components/Navbar/navData.js
@@ -1,15 +1,11 @@
 export const navLinks = [
   {
-    label: 'Projects',
+    label: 'Experience',
     pathname: '/#project-1',
   },
   {
     label: 'Details',
     pathname: '/#details',
-  },
-  {
-    label: 'Articles',
-    pathname: '/articles',
   },
   {
     label: 'Contact',
@@ -19,18 +15,13 @@ export const navLinks = [
 
 export const socialLinks = [
   {
-    label: 'Twitter',
-    url: 'https://twitter.com/hamishMW',
-    icon: 'twitter',
-  },
-  {
-    label: 'Figma',
-    url: 'https://www.figma.com/@Hamish',
-    icon: 'figma',
-  },
-  {
     label: 'Github',
-    url: 'https://github.com/HamishMW',
+    url: 'https://github.com/joaoemersonufc',
     icon: 'github',
+  },
+  {
+    label: 'Linkedin',
+    url: 'https://www.linkedin.com/in/joaoemerson-profissional/',
+    icon: 'linkedin',
   },
 ];
