+++ src/components/ThemeProvider/theme.js
@@ -111,8 +111,8 @@ const dark = {
   themeId: 'dark',
   rgbBackground: '17 17 17',
   rgbBackgroundLight: '26 26 26',
-  rgbPrimary: '0 229 255',
-  rgbAccent: '0 229 255',
+  rgbPrimary: '1 87 255',
+  rgbAccent: '1 87 255',
   rgbText: '255 255 255',
   rgbError: '255 55 102',
   colorTextTitle: 'rgb(var(--rgbText) / 1)',
@@ -125,7 +125,7 @@ const light = {
   rgbBackground: '242 242 242',
   rgbBackgroundLight: '255 255 255',
   rgbPrimary: '0 0 0',
-  rgbAccent: '0 229 255',
+  rgbAccent: '1 87 255',
   rgbText: '0 0 0',
   rgbError: '255 0 60',
   colorTextTitle: 'rgb(var(--rgbText) / 1)',
