+++ src/pages/_app.page.js
@@ -17,22 +17,12 @@ import { ScrollRestore } from '../layouts/App/ScrollRestore';
 
 export const AppContext = createContext({});
 
-const repoPrompt = `
-__  __  __
-\u005C \u005C \u005C \u005C \u005C\u2215\n \u005C \u005C\u2215\u005C \u005C\n  \u005C\u2215  \u005C\u2215
-\n\nTaking a peek huh? Check out the source code: https://github.com/HamishMW/portfolio
-`;
-
 const App = ({ Component, pageProps }) => {
   const [storedTheme] = useLocalStorage('theme', 'dark');
   const [state, dispatch] = useReducer(reducer, initialState);
   const { route, asPath } = useRouter();
   const canonicalRoute = route === '/' ? '' : `${asPath}`;
   useFoucFix();
-
-  useEffect(() => {
-    console.info(`${repoPrompt}\n\n`);
-  }, []);
 
   useEffect(() => {
     dispatch({ type: 'setTheme', value: storedTheme || 'dark' });
