--- src/pages/articles/og-image.css
@@ -1,137 +1,0 @@
-@font-face {
-  font-family: Gotham;
-  font-weight: 400;
-  src: url('../../assets/fonts/gotham-book.woff2') format('woff2');
-  font-display: block;
-  font-style: normal;
-}
-
-@font-face {
-  font-family: Gotham;
-  font-weight: 500;
-  src: url('../../assets/fonts/gotham-medium.woff2') format('woff2');
-  font-display: block;
-  font-style: normal;
-}
-
-@font-face {
-  font-family: Gotham;
-  font-weight: 700;
-  src: url('../../assets/fonts/gotham-bold.woff2') format('woff2');
-  font-display: block;
-  font-style: normal;
-}
-
-:root {
-  --rgbBackground: 17 17 17;
-  --rgbText: 255 255 255;
-  --rgbPrimary: 0 229 255;
-
-  font-family: Gotham, system-ui;
-  background-color: rgb(var(--rgbBackground));
-  color: rgb(var(--rgbText));
-}
-
-.container {
-  background-image: linear-gradient(
-      90deg,
-      rgb(var(--rgbBackground) / 0.95),
-      rgb(var(--rgbBackground) / 0.76)
-    ),
-    var(--url);
-  background-size: cover;
-  background-position: center;
-  position: absolute;
-  inset: 0;
-  display: flex;
-  flex-direction: column;
-  padding: 80px;
-}
-
-.header {
-  display: flex;
-  align-items: center;
-  gap: 24px;
-  margin-top: 32px;
-}
-
-.divider {
-  --lineWidth: 180px;
-  --lineHeight: 4px;
-  --notchWidth: 100px;
-  --notchHeight: 8px;
-
-  position: relative;
-  width: var(--lineWidth);
-  height: var(--lineHeight);
-}
-
-.divider::before,
-.divider::after {
-  content: '';
-  display: block;
-}
-
-.divider::before {
-  width: 100%;
-  height: 100%;
-  background-color: rgb(var(--rgbPrimary));
-}
-
-.divider::after {
-  background-color: rgb(var(--rgbPrimary));
-  position: absolute;
-  clip-path: polygon(0 -1px, 100% -1px, calc(100% - 10px) 100%, 10px 100%);
-  width: var(--notchWidth);
-  height: var(--notchHeight);
-  top: var(--lineHeight);
-}
-
-.date {
-  font-size: calc(24 / 16 * 1rem);
-  font-weight: 500;
-  color: rgb(var(--rgbPrimary));
-}
-
-.title {
-  font-size: calc(86 / 16 * 1rem);
-  line-height: 1.1;
-  letter-spacing: -0.05em;
-  max-width: 80%;
-  margin-top: 48px;
-}
-
-.footer {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-top: auto;
-  color: rgb(var(--rgbText) / 0.7);
-}
-
-.author {
-  display: flex;
-  align-items: center;
-  gap: 24px;
-  font-size: calc(40 / 16 * 1rem);
-  font-weight: 500;
-}
-
-.avatar {
-  clip-path: polygon(0 0, 100% 0, 100% calc(100% - 16px), calc(100% - 16px) 100%, 0 100%);
-}
-
-.timecode {
-  font-size: calc(32 / 16 * 1rem);
-  display: flex;
-  gap: 24px;
-  align-items: center;
-}
-
-.timecode::before {
-  content: '';
-  display: block;
-  height: 4px;
-  width: 100px;
-  background-color: rgb(var(--rgbText) / 0.4);
-}
