+++ src/pages/projects/my-experience/Experience.js
@@ -1,0 +1,339 @@
+import backgroundSprLarge from 'assets/spr-background-large.jpg';
+import backgroundSprPlaceholder from 'assets/spr-background-placeholder.jpg';
+import imageSprBackgroundVolcanismLarge from 'assets/spr-background-volcanism-large.jpg';
+import imageSprBackgroundVolcanismPlaceholder from 'assets/spr-background-volcanism-placeholder.jpg';
+import imageSprBackgroundVolcanism from 'assets/spr-background-volcanism.jpg';
+import backgroundSpr from 'assets/spr-background.jpg';
+import imageSprLessonBuilderDarkLarge from 'assets/spr-lesson-builder-dark-large.jpg';
+import imageSprLessonBuilderDarkPlaceholder from 'assets/spr-lesson-builder-dark-placeholder.jpg';
+import imageSprLessonBuilderDark from 'assets/spr-lesson-builder-dark.jpg';
+import imageSprLessonBuilderLightLarge from 'assets/spr-lesson-builder-light-large.jpg';
+import imageSprLessonBuilderLightPlaceholder from 'assets/spr-lesson-builder-light-placeholder.jpg';
+import imageSprLessonBuilderLight from 'assets/spr-lesson-builder-light.jpg';
+import videoSprMotionLarge from 'assets/spr-motion-large.mp4';
+import videoSprMotionPlaceholder from 'assets/spr-motion-placeholder.jpg';
+import videoSprMotion from 'assets/spr-motion.mp4';
+import imageSprStoryboarderDarkLarge from 'assets/spr-storyboarder-dark-large.png';
+import imageSprStoryboarderDarkPlaceholder from 'assets/spr-storyboarder-dark-placeholder.png';
+import imageSprStoryboarderDark from 'assets/spr-storyboarder-dark.png';
+import imageSprStoryboarderLightLarge from 'assets/spr-storyboarder-light-large.png';
+import imageSprStoryboarderLightPlaceholder from 'assets/spr-storyboarder-light-placeholder.png';
+import imageSprStoryboarderLight from 'assets/spr-storyboarder-light.png';
+import { Footer } from 'components/Footer';
+import { Image } from 'components/Image';
+import { Meta } from 'components/Meta';
+import { ThemeProvider, useTheme } from 'components/ThemeProvider';
+import { Button } from 'components/Button';
+import {
+  ProjectBackground,
+  ProjectContainer,
+  ProjectHeader,
+  ProjectImage,
+  ProjectSection,
+  ProjectSectionColumns,
+  ProjectSectionContent,
+  ProjectSectionHeading,
+  ProjectSectionText,
+  ProjectTextRow,
+} from 'layouts/Project';
+import dynamic from 'next/dynamic';
+import { Fragment, useMemo } from 'react';
+import { media } from 'utils/style';
+import styles from './Experience.module.css';
+
+const Earth = dynamic(() => import('./Earth').then(mod => mod.Earth));
+const EarthSection = dynamic(() => import('./Earth').then(mod => mod.EarthSection));
+
+const title = 'More than 5 years of work';
+const description =
+  'I love to learning new features and frameworks and apply to the next generation website designs. In the right side you can see some areas thats I already worked at';
+const roles = [
+  'Innovation',
+  'Logistics',
+  'Telecommunications',
+  'Health',
+  'Pharmaceuticals',
+  'General Software Development',
+];
+
+export const Experience = () => {
+  const { themeId } = useTheme();
+
+  const isDark = themeId === 'dark';
+
+  return (
+    <Fragment>
+      <ProjectContainer className="spr">
+        <Meta title={title} prefix="Projects" description={description} />
+        <ProjectBackground
+          opacity={isDark ? 0.5 : 0.8}
+          src={backgroundSpr}
+          srcSet={`${backgroundSpr.src} 1080w, ${backgroundSprLarge.src} 2160w`}
+          placeholder={backgroundSprPlaceholder}
+        />
+        <ProjectHeader
+          title={title}
+          description={description}
+          url="https://bit.ly/joaoemersonengineer"
+          roles={roles}
+        />
+        <ProjectSection padding="top">
+          <ProjectSectionContent>
+            <ProjectImage
+              raised
+              key={themeId}
+              srcSet={
+                isDark
+                  ? [imageSprLessonBuilderDark, imageSprLessonBuilderDarkLarge]
+                  : [imageSprLessonBuilderLight, imageSprLessonBuilderLightLarge]
+              }
+              placeholder={
+                isDark
+                  ? imageSprLessonBuilderDarkPlaceholder
+                  : imageSprLessonBuilderLightPlaceholder
+              }
+              sizes={`(max-width: ${media.mobile}px) 100vw, (max-width: ${media.tablet}px) 800px, 1000px`}
+              alt="The aero lesson builder app dragging an audio component into a screen about plant cells."
+            />
+          </ProjectSectionContent>
+        </ProjectSection>
+        <ProjectSection>
+          <ProjectTextRow>
+            <ProjectSectionHeading>The things change</ProjectSectionHeading>
+            <ProjectSectionText>
+            We all need to imagine how we can change, improve and evolve.
+            I did the same by transforming my old portfolio into pure HTML in this one with the most current technologies on the market
+            </ProjectSectionText>
+          </ProjectTextRow>
+        </ProjectSection>
+        <ThemeProvider themeId="dark" data-invert>
+          <ProjectSection
+            backgroundOverlayOpacity={0.5}
+            backgroundElement={
+              <Image
+                srcSet={[imageSprBackgroundVolcanism, imageSprBackgroundVolcanismLarge]}
+                placeholder={imageSprBackgroundVolcanismPlaceholder}
+                alt="A dramatic ocean scene with lava forming a new land mass."
+                sizes="100vw"
+              />
+            }
+          >
+            <ProjectSectionColumns width="full">
+              <ProjectSectionContent width="full">
+                <ProjectTextRow width="s">
+                  <ProjectSectionHeading>See my Linkedin Profile</ProjectSectionHeading>
+                  <ProjectSectionText>
+                    If you are interested in my work, get in touch using the menu on the left.
+                    Or if you prefer, visit my Linkedin page, I'm grateful that you've made it this far.
+                  </ProjectSectionText>
+                </ProjectTextRow>
+                <Button
+                  primary
+                  iconHoverShift
+                  className={styles.linkButton}
+                  icon="chevronRight"
+                  href={'https://www.linkedin.com/in/joaoemerson-profissional/'}
+                >
+                  {'Visit the Linkedin profile'}
+                </Button>
+              </ProjectSectionContent>
+              <Image
+                raised
+                className={styles.video}
+                srcSet={[
+                  { src: videoSprMotion, width: 1280 },
+                  { src: videoSprMotionLarge, width: 2560 },
+                ]}
+                placeholder={videoSprMotionPlaceholder}
+                alt="A small demonstration of my Linkedin Profile"
+                sizes={`(max-width: ${media.mobile}px) 100vw, 50vw`}
+              />
+            </ProjectSectionColumns>
+          </ProjectSection>
+        </ThemeProvider>
+        <ProjectSection>
+          <ProjectSectionContent>
+            <ProjectTextRow>
+              <ProjectSectionHeading>Encouraging adaptivity</ProjectSectionHeading>
+              <ProjectSectionText>
+              I have a lot of experience with project and team management tools like JIRA, Trello, GitHub, Notion.
+              In addition to having experience with various communication channels such as Slack, Skype, Microsoft Teams, Google Chats and etc.
+              </ProjectSectionText>
+            </ProjectTextRow>
+            <Image
+              raised
+              key={themeId}
+              srcSet={
+                isDark
+                  ? [imageSprStoryboarderDark, imageSprStoryboarderDarkLarge]
+                  : [imageSprStoryboarderLight, imageSprStoryboarderLightLarge]
+              }
+              placeholder={
+                isDark
+                  ? imageSprStoryboarderDarkPlaceholder
+                  : imageSprStoryboarderLightPlaceholder
+              }
+              alt="A drag and drop storyboard style editor for creating an adaptive lesson."
+              sizes={`(max-width: ${media.mobile}px) 100vw, 80vw`}
+            />
+          </ProjectSectionContent>
+        </ProjectSection>
+        <ProjectSection>
+          <ProjectSectionColumns>
+            <ProjectSectionContent>
+              <ProjectTextRow>
+                <ProjectSectionHeading>
+                  Its just a beautiful explanation of how javascript is powerful. See below.
+                </ProjectSectionHeading>
+              </ProjectTextRow>
+            </ProjectSectionContent>
+          </ProjectSectionColumns>
+        </ProjectSection>
+        <ThemeProvider themeId="dark" data-invert>
+          <Earth
+            className={styles.earth}
+            hideMeshes={useMemo(
+              () => ['Atmosphere', 'EarthPartial', 'Chunk', 'EarthFull'],
+              []
+            )}
+            position={useMemo(() => [0, 0, 0], [])}
+            labels={useMemo(
+              () => [
+                {
+                  position: [0.54, 0.19, 0.18],
+                  text: 'Pacific ring of fire',
+                  hidden: true,
+                },
+                {
+                  position: [0.47, -0.38, 0.04],
+                  text: 'Ruapehu',
+                  hidden: true,
+                },
+                {
+                  position: [0.22, 0.44, -0.35],
+                  text: 'St. Helens',
+                  hidden: true,
+                },
+                {
+                  position: [0.16, -0.06, 0.58],
+                  text: 'Krakatoa',
+                  hidden: true,
+                },
+                {
+                  position: [0.11, 0.2, -0.56],
+                  text: 'Parícutin',
+                  hidden: true,
+                },
+                {
+                  position: [0.52, 0.2, -0.23],
+                  text: 'Kīlauea',
+                  hidden: true,
+                },
+                {
+                  position: [-0.24, 0.75, 0.24],
+                  text: 'Mantle',
+                  delay: 800,
+                  hidden: true,
+                },
+                {
+                  position: [-0.24, 0.55, 0.24],
+                  text: 'Outer core',
+                  delay: 800,
+                  hidden: true,
+                },
+                {
+                  position: [-0.24, 0.35, 0.24],
+                  text: 'Inner core',
+                  delay: 800,
+                  hidden: true,
+                },
+              ],
+              []
+            )}
+            scale={0.6}
+          >
+            <EarthSection
+              scrim
+              animations={['0:loop']}
+              camera={[0, 0, 1.5]}
+              meshes={['Atmosphere', 'EarthFull']}
+            >
+              <ProjectSection>
+                <ProjectSectionContent>
+                  <ProjectTextRow center>
+                    <ProjectSectionHeading>
+                      Next-generation learning experiences
+                    </ProjectSectionHeading>
+                  </ProjectTextRow>
+                </ProjectSectionContent>
+              </ProjectSection>
+            </EarthSection>
+            <EarthSection
+              animations={['0:loop']}
+              camera={[0, 0, 2.4]}
+              meshes={['Atmosphere', 'EarthFull']}
+            />
+            <EarthSection
+              animations={['0:loop']}
+              camera={[1.14, -1.39, 0.94]}
+              meshes={['Atmosphere', 'EarthFull']}
+            >
+              <ProjectSection>
+                <ProjectSectionContent width="xl">
+                  <ProjectTextRow justify="end" width="s">
+                    <ProjectSectionHeading level={4} as="h3">
+                      Bringing 3D into learning
+                    </ProjectSectionHeading>
+                  </ProjectTextRow>
+                </ProjectSectionContent>
+              </ProjectSection>
+            </EarthSection>
+            <EarthSection
+              animations={['0:loop']}
+              camera={[1.17, 0.69, -1.47]}
+              meshes={['Atmosphere', 'EarthFull']}
+              labels={[
+                'Pacific ring of fire',
+                'Ruapehu',
+                'St. Helens',
+                'Krakatoa',
+                'Parícutin',
+                'Kīlauea',
+              ]}
+            >
+              <ProjectSection>
+                <ProjectSectionContent width="xl">
+                  <ProjectTextRow justify="start" width="s">
+                    <ProjectSectionHeading level={4} as="h3">
+                      Interactivity
+                    </ProjectSectionHeading>
+                  </ProjectTextRow>
+                </ProjectSectionContent>
+              </ProjectSection>
+            </EarthSection>
+            <EarthSection
+              animations={['0:loop']}
+              camera={[1.81, 0.51, 0.43]}
+              meshes={['Atmosphere', 'EarthFull']}
+              labels={[
+                'Pacific ring of fire',
+                'Ruapehu',
+                'St. Helens',
+                'Krakatoa',
+                'Parícutin',
+                'Kīlauea',
+              ]}
+            />
+            <EarthSection
+              scrimReverse
+              animations={['0:loop']}
+              camera={[0.37, 1.02, 1.84]}
+              meshes={['Atmosphere', 'EarthFull']}
+            />
+          </Earth>
+        </ThemeProvider>
+      </ProjectContainer>
+      <Footer />
+    </Fragment>
+  );
+};
