+++ src/pages/projects/my-experience/index.page.js
@@ -1,0 +1,1 @@
+export { Experience as default } from './Experience';
