--- src/pages/projects/smart-sparrow/index.page.js
@@ -1,1 +1,0 @@
-export { SmartSparrow as default } from './SmartSparrow';
